Fixed failing junits.
