+++ src/main/java/hudson/scm/subversion/CheckoutUpdater.java
@@ -137,4 +137,20 @@ public class CheckoutUpdater extends WorkspaceUpdater {
             Util.deleteContentsRecursive(ws);
         }
     }
+    
+    @Override
+    public boolean equals(Object o) {
+    	if (this == o)
+    		return true;
+    	
+    	if (!(o instanceof CheckoutUpdater))
+    		return false;
+    	return true;
+    }
+    
+    @Override
+    public int hashCode() {
+    	return getClass().hashCode();
+    	
+    }
 }
